Addressed feedback regarding sortJSON and DSNP types
