+++ src/social/search.ts
@@ -1,6 +1,6 @@
 import { Handle } from "./handles";
 import { MessageType } from "../types/DSNP";
-import { HexString } from "../types/String";
+import { HexString } from "../types/Strings";
 import { NotImplementedError } from "../utilities/errors";
 
 type BlockNumber = number;
